requirements.txt
instructor>=1.3.0
openai>=1.40.0
pydantic>=2.7.0
Pillow>=10.0.0
pyautogui>=0.9.54
pywin32>=306
keyboard>=0.13.5
ultralytics>=8.3.0
opencv-python>=4.10.0.84
numpy>=1.26.0
pytesseract>=0.3.10
